Move v2 apply logic into apply directory
